+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.21.0 [unreleased]
 
+### Features
+1. [#xxx](https://github.com/influxdata/influxdb-client-python/pull/xxx): Add supports for array expressions in query parameters
+ 
 ## 1.20.0 [2021-08-20]
 
 ### Features